requirements.txt
utils-nuuuwan
gig-nuuuwan
numpy
pandas
matplotlib
Pillow